Add elasticsearch/kibana.
Logs aren't indexed during the run of an experiment.
This commit is about indexing all the logs in the post-mortem phase.
To do so I reuse the heka docker image from kolla and all the files
allowing the pipeline between the log and elasticsearch to work.

In comparision to the original decoders we allow the customization of
the Type field to differentiate logs from different experiments.
Note that the indexation process may take some time !

M